// src/blueprints/tv2/enums/tv2-piece-type.ts
export enum Tv2PieceType {
  CAMERA = 'CAMERA',
  REPLAY = 'REPLAY',
  REMOTE = 'REMOTE',
  GRAPHICS = 'GRAPHICS',
  OVERLAY_GRAPHICS = 'OVERLAY_GRAPHICS',
  SPLIT_SCREEN = 'SPLIT_SCREEN',
  VIDEO_CLIP = 'VIDEO_CLIP',
  VOICE_OVER = 'VOICE_OVER',
  JINGLE = 'JINGLE',
  AUDIO = 'AUDIO',
  MANUS = 'MANUS',
  TRANSITION = 'TRANSITION',
  COMMAND = 'COMMAND',
  UNKNOWN = 'UNKNOWN',
}